+++ src/scss/scaffold.scss
@@ -95,79 +95,6 @@ body {
   .widgetPanel.disabled {
     display: none;
   }
-
-  .preview-controller-wrap {
-    display :none;
-    &.active {
-      display: block;
-    }
-  }
-
-  .interaction-preview-container .preview-controller {
-    background: $lgrey-bg;
-    width: 100%;
-    display: none;
-
-    &.active {
-      display: block;
-    }
-
-    h2 {
-      justify-content: normal;
-      margin: 0 $outer-padding;
-    }
-
-    h5  {
-      margin: $outer-padding $outer-padding $outer-padding / 2;
-    }
-
-    .preview-close {
-      margin-left: auto;
-      margin-right: $outer-padding;
-      cursor: pointer;
-    }
-
-    .preview-header {
-      display: flex;
-      align-items: center;
-      padding: $inner-padding 0;
-      background: $dgrey-bg;
-      color: white;
-      fill: white;
-    }
-
-    .preview-header h2 {
-      display: inline;
-    }
-
-    .preview-scroll {
-      display: flex;
-      background: $grey-bg;
-    }
-
-    .preview-scroll > div {
-      display: inline-block;
-      padding: 0 $outer-padding $outer-padding;
-      background: $grey-bg;
-      color: white;
-
-      &.selected {
-        background: $lgrey-bg;
-      }
-
-      .preview-label {
-        margin: 0.5 * $outer-padding 0;
-      }
-
-      svg.marks > g {
-        transform: translate(12px, 13px);
-      }
-    }
-
-    .signals-container {
-      margin-left: 10px;
-    }
-  }
 }
 
 // each vertical sidebar
